pyproject.toml
[project]
name = "finals-project-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.5",
    "python-dotenv>=1.1.1",
    "tensorflow>=2.20.0",
]
